refactor(leansig): move leansig packages under xmss/ and fix comments
