+++ test/unit/forkchoice_attestation_validation_test.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/geanlabs/gean/chain/forkchoice"
 	"github.com/geanlabs/gean/chain/statetransition"
-	"github.com/geanlabs/gean/leansig"
+	"github.com/geanlabs/gean/xmss/leansig"
 	"github.com/geanlabs/gean/types"
 )
