+++ xmss/leansig-ffi/include/leansig_ffi.h
